requirements.txt
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
flask-restx==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python-headless==4.2.0.34
Pillow==7.1.2
pyrsistent==0.16.0
pytesseract==0.3.4
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
gunicorn 
